a bit of a mess right now but committing for public viewing
